QuickAdminPanel automatic commit

+++ database/migrations/2020_11_20_000002_create_audit_logs_table.php
@@ -9,11 +9,11 @@ class CreateAuditLogsTable extends Migration
     public function up()
     {
         Schema::create('audit_logs', function (Blueprint $table) {
-            $table->increments('id');
+            $table->bigIncrements('id');
             $table->text('description');
-            $table->unsignedInteger('subject_id')->nullable();
+            $table->unsignedBigInteger('subject_id')->nullable();
             $table->string('subject_type')->nullable();
-            $table->unsignedInteger('user_id')->nullable();
+            $table->unsignedBigInteger('user_id')->nullable();
             $table->text('properties')->nullable();
             $table->string('host', 45)->nullable();
             $table->timestamps();
